modify status updation and cleanup resource naming convention
