+++ controllers/argocd/appcontroller/status.go
@@ -1,12 +1,10 @@
 package appcontroller
 
 import (
-	"context"
-
 	"github.com/argoproj-labs/argocd-operator/common"
+	"github.com/argoproj-labs/argocd-operator/pkg/resource"
 	"github.com/argoproj-labs/argocd-operator/pkg/workloads"
 	"github.com/pkg/errors"
-	"k8s.io/client-go/util/retry"
 )
 
 // reconcileStatus will ensure that the app-controller status is updated for the given ArgoCD instance
@@ -36,17 +34,5 @@ func (acr *AppControllerReconciler) ReconcileStatus() error {
 }
 
 func (acr *AppControllerReconciler) updateInstanceStatus() error {
-	err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
-		if err := acr.Client.Status().Update(context.TODO(), acr.Instance); err != nil {
-			return errors.Wrap(err, "UpdateInstanceStatus: failed to update instance status")
-		}
-		return nil
-	})
-
-	if err != nil {
-		// May be conflict if max retries were hit, or may be something unrelated
-		// like permissions or a network error
-		return err
-	}
-	return nil
+	return resource.UpdateStatusSubResource(acr.Instance, acr.Client)
 }
