--- controllers/argocd/deployment.go
@@ -1,140 +1,0 @@
-// Copyright 2019 ArgoCD Operator Developers
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// 	http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package argocd
-
-import (
-	"fmt"
-	"os"
-	"reflect"
-	"strings"
-
-	argoproj "github.com/argoproj-labs/argocd-operator/api/v1beta1"
-	"github.com/argoproj-labs/argocd-operator/common"
-	"github.com/argoproj-labs/argocd-operator/pkg/argoutil"
-
-	appsv1 "k8s.io/api/apps/v1"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-)
-
-// getDexServerAddress will return the Dex server address.
-func getDexServerAddress(cr *argoproj.ArgoCD) string {
-	return fmt.Sprintf("https://%s", fqdnServiceRef("dex-server", common.ArgoCDDefaultDexHTTPPort, cr))
-}
-
-// newDeployment returns a new Deployment instance for the given ArgoCD.
-func newDeployment(cr *argoproj.ArgoCD) *appsv1.Deployment {
-	return &appsv1.Deployment{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      cr.Name,
-			Namespace: cr.Namespace,
-			Labels:    argoutil.LabelsForCluster(cr),
-		},
-	}
-}
-
-// newDeploymentWithName returns a new Deployment instance for the given ArgoCD using the given name.
-func newDeploymentWithName(name string, component string, cr *argoproj.ArgoCD) *appsv1.Deployment {
-	deploy := newDeployment(cr)
-	deploy.ObjectMeta.Name = name
-
-	lbls := deploy.ObjectMeta.Labels
-	lbls[common.ArgoCDKeyName] = name
-	lbls[common.ArgoCDKeyComponent] = component
-	deploy.ObjectMeta.Labels = lbls
-
-	deploy.Spec = appsv1.DeploymentSpec{
-		Selector: &metav1.LabelSelector{
-			MatchLabels: map[string]string{
-				common.ArgoCDKeyName: name,
-			},
-		},
-		Template: corev1.PodTemplateSpec{
-			ObjectMeta: metav1.ObjectMeta{
-				Labels: map[string]string{
-					common.ArgoCDKeyName: name,
-				},
-			},
-			Spec: corev1.PodSpec{
-				NodeSelector: common.DefaultNodeSelector(),
-			},
-		},
-	}
-
-	if cr.Spec.NodePlacement != nil {
-		deploy.Spec.Template.Spec.NodeSelector = argoutil.AppendStringMap(deploy.Spec.Template.Spec.NodeSelector, cr.Spec.NodePlacement.NodeSelector)
-		deploy.Spec.Template.Spec.Tolerations = cr.Spec.NodePlacement.Tolerations
-	}
-	return deploy
-}
-
-// newDeploymentWithSuffix returns a new Deployment instance for the given ArgoCD using the given suffix.
-func newDeploymentWithSuffix(suffix string, component string, cr *argoproj.ArgoCD) *appsv1.Deployment {
-	return newDeploymentWithName(fmt.Sprintf("%s-%s", cr.Name, suffix), component, cr)
-}
-
-// reconcileDeployments will ensure that all Deployment resources are present for the given ArgoCD.
-func (r *ReconcileArgoCD) reconcileDeployments(cr *argoproj.ArgoCD, useTLSForRedis bool) error {
-
-	if err := r.reconcileDexDeployment(cr); err != nil {
-		log.Error(err, "error reconciling dex deployment")
-	}
-
-	err := r.reconcileRedisDeployment(cr, useTLSForRedis)
-	if err != nil {
-		return err
-	}
-
-	err = r.reconcileRedisHAProxyDeployment(cr)
-	if err != nil {
-		return err
-	}
-
-	err = r.reconcileRepoDeployment(cr, useTLSForRedis)
-	if err != nil {
-		return err
-	}
-
-	err = r.reconcileServerDeployment(cr, useTLSForRedis)
-	if err != nil {
-		return err
-	}
-
-	err = r.reconcileGrafanaDeployment(cr)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func isRemoveManagedByLabelOnArgoCDDeletion() bool {
-	if v := os.Getenv("REMOVE_MANAGED_BY_LABEL_ON_ARGOCD_DELETION"); v != "" {
-		return strings.ToLower(v) == "true"
-	}
-	return false
-}
-
-// to update nodeSelector and tolerations in reconciler
-func updateNodePlacement(existing *appsv1.Deployment, deploy *appsv1.Deployment, changed *bool) {
-	if !reflect.DeepEqual(existing.Spec.Template.Spec.NodeSelector, deploy.Spec.Template.Spec.NodeSelector) {
-		existing.Spec.Template.Spec.NodeSelector = deploy.Spec.Template.Spec.NodeSelector
-		*changed = true
-	}
-	if !reflect.DeepEqual(existing.Spec.Template.Spec.Tolerations, deploy.Spec.Template.Spec.Tolerations) {
-		existing.Spec.Template.Spec.Tolerations = deploy.Spec.Template.Spec.Tolerations
-		*changed = true
-	}
-}
