+++ controllers/argocd/notifications/status.go
@@ -1,11 +1,6 @@
 package notifications
 
-import (
-	"context"
-
-	"github.com/pkg/errors"
-	"k8s.io/client-go/util/retry"
-)
+import "github.com/argoproj-labs/argocd-operator/pkg/resource"
 
 // reconcileStatus will ensure that the notifications controller status is updated for the given ArgoCD instance
 func (nr *NotificationsReconciler) ReconcileStatus() error {
@@ -16,17 +11,5 @@ func (nr *NotificationsReconciler) ReconcileStatus() error {
 }
 
 func (nr *NotificationsReconciler) updateInstanceStatus() error {
-	err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
-		if err := nr.Client.Status().Update(context.TODO(), nr.Instance); err != nil {
-			return errors.Wrap(err, "UpdateInstanceStatus: failed to update instance status")
-		}
-		return nil
-	})
-
-	if err != nil {
-		// May be conflict if max retries were hit, or may be something unrelated
-		// like permissions or a network error
-		return err
-	}
-	return nil
+	return resource.UpdateStatusSubResource(nr.Instance, nr.Client)
 }
