+++ controllers/argocd/redis/status.go
@@ -1,12 +1,10 @@
 package redis
 
 import (
-	"context"
-
 	"github.com/argoproj-labs/argocd-operator/common"
+	"github.com/argoproj-labs/argocd-operator/pkg/resource"
 	"github.com/argoproj-labs/argocd-operator/pkg/workloads"
 	"github.com/pkg/errors"
-	"k8s.io/client-go/util/retry"
 )
 
 // ReconcileStatus will ensure that the Redis status is updated for the given ArgoCD instance
@@ -51,17 +49,5 @@ func (rr *RedisReconciler) ReconcileStatus() error {
 }
 
 func (rr *RedisReconciler) updateInstanceStatus() error {
-	err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
-		if err := rr.Client.Status().Update(context.TODO(), rr.Instance); err != nil {
-			return errors.Wrap(err, "UpdateInstanceStatus: failed to update instance status")
-		}
-		return nil
-	})
-
-	if err != nil {
-		// May be conflict if max retries were hit, or may be something unrelated
-		// like permissions or a network error
-		return err
-	}
-	return nil
+	return resource.UpdateStatusSubResource(rr.Instance, rr.Client)
 }
