+++ controllers/argocd/reposerver/status.go
@@ -1,14 +1,12 @@
 package reposerver
 
 import (
-	"context"
-
 	"github.com/argoproj-labs/argocd-operator/common"
+	"github.com/argoproj-labs/argocd-operator/pkg/resource"
 	"github.com/argoproj-labs/argocd-operator/pkg/workloads"
 	"github.com/pkg/errors"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/client-go/util/retry"
 )
 
 // ReconcileStatus will ensure that the Repo-server status is updated for the given ArgoCD instance
@@ -46,17 +44,5 @@ func (rsr *RepoServerReconciler) ReconcileStatus() error {
 }
 
 func (rsr *RepoServerReconciler) updateInstanceStatus() error {
-	err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
-		if err := rsr.Client.Status().Update(context.TODO(), rsr.Instance); err != nil {
-			return errors.Wrap(err, "UpdateInstanceStatus: failed to update instance status")
-		}
-		return nil
-	})
-
-	if err != nil {
-		// May be conflict if max retries were hit, or may be something unrelated
-		// like permissions or a network error
-		return err
-	}
-	return nil
+	return resource.UpdateStatusSubResource(rsr.Instance, rsr.Client)
 }
