+++ controllers/argocd/sso/status.go
@@ -1,11 +1,8 @@
 package sso
 
 import (
-	"context"
-
 	argoproj "github.com/argoproj-labs/argocd-operator/api/v1beta1"
-	"github.com/pkg/errors"
-	"k8s.io/client-go/util/retry"
+	"github.com/argoproj-labs/argocd-operator/pkg/resource"
 )
 
 // reconcileStatus will ensure that the sso status is updated for the given ArgoCD instance
@@ -33,17 +30,5 @@ func (sr *SSOReconciler) ReconcileStatus() error {
 }
 
 func (sr *SSOReconciler) updateInstanceStatus() error {
-	err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
-		if err := sr.Client.Status().Update(context.TODO(), sr.Instance); err != nil {
-			return errors.Wrap(err, "UpdateInstanceStatus: failed to update instance status")
-		}
-		return nil
-	})
-
-	if err != nil {
-		// May be conflict if max retries were hit, or may be something unrelated
-		// like permissions or a network error
-		return err
-	}
-	return nil
+	return resource.UpdateStatusSubResource(sr.Instance, sr.Client)
 }
